Downgrade SQLite upserts to 'update if exists' + add 3.9-dev to travis vers
Unfortunately, SQLite upserts (`ON CONFLICT(name) DO UPDATE xxx`) are a rather new feature (released as of SQLite 3.24.0 2018-06-04),
and it appears that the latest SQLite3 version in Ubuntu 18.04 (Bionic)'s repos is 3.22.0 - meaning that upserts are likely to be unusable
on many systems.

Due to this issue, upsert statements in `SqliteCacheManager` and `AsyncSqliteCacheManager` have been replaced with standard INSERT statements,
and the `update_cache_key` method (in both sync and async classes) now uses a standard `UPDATE` statement, which should work on all systems.

Additionally, `3.9-dev` has now been added to `.travis.yml`, but marked with `allow_failures` to avoid bugs in 3.9 (or problems with third-party libraries
that haven't yet been patched to work with 3.9) causing the travis build to fail.